rag/app/requirements.txt
fastapi
uvicorn[standard]
pillow

langchain
langchain-community
langchain-core
langchain-mongodb
pymongo
langgraph
pymupdf4llm
ocrmypdf
fitz
rank_bm25